Move store out of containers.
